refactor: zero payload 방식으로 이벤트 dto 리팩토링
