+++ notification-application/src/main/kotlin/gloddy/notification/dto/event/ApplyEvent.kt
@@ -1,11 +1,10 @@
-package gloddy.notification.dto
+package gloddy.notification.dto.event
 
-import gloddy.notification.dto.eventType.ApplyEventType
-import gloddy.notification.UserId
+import gloddy.notification.dto.event.eventType.ApplyEventType
 import java.time.LocalDateTime
 
-data class ApplyEvent(
-    val userId: UserId,
+data class ApplyEvent (
+    val applyId: Long,
     val eventType: ApplyEventType,
     val eventDateTime: LocalDateTime
 )
