+++ notification-application/src/main/kotlin/gloddy/notification/dto/event/GroupMemberEvent.kt
@@ -1,9 +1,10 @@
-package gloddy.notification.dto
+package gloddy.notification.dto.event
 
-import gloddy.notification.dto.eventType.GroupMemberEventType
+import gloddy.notification.dto.event.eventType.GroupMemberEventType
+import java.time.LocalDateTime
 
 data class GroupMemberEvent(
     val groupMemberId: Long,
     val eventType: GroupMemberEventType,
-    val 
+    val eventDateTime: LocalDateTime
 )
