+++ notification-application/src/main/kotlin/gloddy/notification/dto/event/eventType/ApplyEventType.kt
@@ -1,4 +1,4 @@
-package gloddy.notification.dto.eventType
+package gloddy.notification.dto.event.eventType
 
 enum class ApplyEventType {
     APPLY_APPROVE,
