+++ notification-application/src/main/kotlin/gloddy/notification/dto/event/eventType/GroupEventType.kt
@@ -1,4 +1,4 @@
-package gloddy.notification.dto.eventType
+package gloddy.notification.dto.event.eventType
 
 enum class GroupEventType {
     GROUP_APPROACHING_START,
