+++ notification-application/src/main/kotlin/gloddy/notification/dto/event/eventType/GroupMemberEventType.kt
@@ -1,4 +1,4 @@
-package gloddy.notification.dto.eventType
+package gloddy.notification.dto.event.eventType
 
 enum class GroupMemberEventType {
     GROUP_LEAVE
